+++ system/src/Grav/Framework/Flex/FlexDirectoryForm.php
@@ -130,6 +130,17 @@ class FlexDirectoryForm implements FlexDirectoryFormInterface, JsonSerializable
     }
 
     /**
+     * @param string $uniqueId
+     * @return void
+     */
+    public function setUniqueId(string $uniqueId): void
+    {
+        if ($uniqueId !== '') {
+            $this->uniqueid = $uniqueId;
+        }
+    }
+
+    /**
      * @param string $name
      * @param mixed $default
      * @param string|null $separator